requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ultralytics>=8.0.0
sentence-transformers>=2.2.2
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
imagehash>=4.3.1
clip @ git+https://github.com/openai/CLIP.git

# Network Graph Libraries
networkx>=3.1
neo4j>=5.14.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Google AI
google-generativeai>=0.3.0

# Firebase
firebase-admin>=6.0.0

# Supabase
supabase>=2.0.0

# Groq AI (fallback for rate limiting)
groq>=0.4.0

# Phase 2: XAI & Explainability
shap>=0.43.0
lime>=0.2.0.1
captum>=0.6.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# OCR
easyocr>=1.7.1

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2
pytz>=2023.3

# Monitoring & Metrics
prometheus-client>=0.18.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Data Processing
pandas>=2.0.0

# Training & Visualization
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
